Create CommandHistory TestCases
